requirements.txt
python-telegram-bot==13.7
aiohttp==3.9.1
python-dotenv==1.0.0
asyncio==3.4.3
typing==3.7.4.3
python-dateutil==2.8.2
Pillow==11.2.1 